PowerPC: Clear floating-point exceptions before enabling them
In double-precision PowerPC builds with GCC 9.4 and 11.1, the call to
feenableexcept triggers an FPE if there have been an illegal operation
before the call. Such situation happens in MdrunNonIntegratorTests,
although it should not happen during normal md runs.

To avoid the problem, we now explicitly clear the FP exception flags.