Make MPI build support conditional on user choices.
Various build options relied on an assumption that `find_package(MPI)`
would not be called (and output variables would not be set) if not
needed for a particular target.

This change decouples the use of `FindMPI.cmake` from target dependence
on its results.

Deferred:
There is still a lot of overuse of global CMake environment, and
reliance on deprecated CMake infrastructure.

See also #3672

Closes #4678