Suppress OpenMP thread count warning for DD GPU runs
For runs with DD, we get the warning that the optimum number of OpenMP
threads is "usually between 2 and 8 threads per rank." But this is
misleading for GPU runs, where e.g. on a 4-GPU server with 2xRome CPUs
(128 cores) we often run with 4 MPI tasks and the best thread count
per MPI task is 32. This patch supresses this warning if GPUs are in
use.