Support HeFFTe in SYCL build
In SYCL+HeFFTe configurations, the backend is hardcoded to oneMKL or
rocFFT in the natural way.

Converted PME grid halo-exchange GPU kernels to SYCL. The logic in the
kernels is unchanged. De-duplicated the kernel-launch logic compared
with the original CUDA implementation.  Reordered kernel arguments to
permit template parameter packs to reduce verbosity. Used size_t
instead of int for compatibility. Used more DeviceBuffer and thus
asMpiPointer as required. Slightly improved the doxygen.

Allowed construction of DeviceAccessor from empty DeviceBuffers. The
grid halo-exchange kernel takes responsibility for not using such
accessors when there is no exchange in that dimension.

Refs #4153, #4090