Add new AUTHORS
This is not a comprehensive update.

Based on `git log --since="2022-01-01" --format='%an'` and
https://gitlab.com/groups/gromacs/online-tutorials/-/activity.

Added people with at least two contributions or a single big one.
Have not moved anyone to "previous contributors".

Sorry if someone was forgotten here, we plan to update the author
list once more before 2023 release.

Refs #4664