Support double-batched FFT library
This open-source library from MKL provides a more even baseline of
high performance than MKL, which achieves high performance only for
specifically tuned problem sizes.

It was formerly known as bbfft. The public-facing name was changed to
suit open-sourcing requirements. However, the source code continues to
use bbfft for backwards compatibility and the GROMACS implementation
here shares that history.

Only an external library is supported. There is no need to bundle one
at this time, because few GROMACS users will need it, and no distros
need to package with it. The source itself needs minor improvements to
be a nestable CMake project if we want that in future.

Refs #4153