Also use a cut-off for FEP lists with CPU kernels
As when using GPU kernels, the FEP pairlist with CPU kernels now also
uses a cut-off of rlist plus a 1x1 buffer distance instead of putting
all atom pairs in cluster pairs in the FEP list. This harmonizes
the code and should also give a small performance improvement, but more
importantly it avoids double corrections for excluded pairs in systems
with few atoms, issue #4665. Note that this issue was fixed in the FEP
non-bonded kernel instead in release-2022.

Fixes #4665