Disable CUDA graph earlier whenever possible
CUDA graph scheduling is not compatible when CPU force tasks are
present, but since this information is not readily available at the time
simulationWorkload.useMdGpuGraph is set, we left the determination of
whether CUDA graph is actually used for launching tasks
to a later stage based on DomainLifetimeWorkload flag.
This commit adds a few additional conditions to the initialization of
useMdGpuGraph so at least in the cases that can be identified early we
avoid reporting that CUDA graph may be used.

Refs #4277