Adjust logging output for pytest test fixture
"debug" level output is suppressed in the default logging configuration. When we know there is an error, it makes sense to log `stderr` to the "error" level for more useful terminal output during pytest.

`stdout` is probably less useful, but we might as well make it available for developers who explicitly configure "debug" level pytest console logging.