Allow selection of energy term by full name
When the energy term name starts with a number, like "1\Viscosity", entering
this name at the prompt would select term 1. This fixes this by first searching
for a full exact match of the name. (then, by number, and finally by start
of the name, as previously)

Adds a test for this.

**Supported**: Select by full name, then select by number or partial name on another lines

**Unsupported**: Select by full name, and by number or partial name on the same line.
Selection by partial name for name starting with a number.

Fixes #4739