Remove a superfluous documentation section.
The gmxapi documentation section is a strange place to see links to the
top level index and search pages. This was a leftover section from when
gmxapi Python package documentation could be decoupled from the main
documentation build. We don't do that anymore, so the section had become
confusing (worse than useless).