Remove usage of MKLConfig
Fixes #4706 

I created this to see how much code it would be if we don't rely on either MKLConfig or -qmkl. Goal was to keep it as simple as possible. It turns out that without static library support it isn't more code than using MKLConfig. But adding support for static linking about doubles the required lines.

RFC because:
1) Do we want this solution over the other option proposed in #4706 
2) If so do we want the support for static linking (2nd commit) or the simpler version without the support.

Note that there is anyhow libsycl.a and therefore it isn't possible to have a fully static binary with DPC++. What is possible is to have MKL statically linked and a fully static non SYCL binary. 

The `find_library` change in the 2nd commit isn't needed to support fully static binary. It is needed if we want `GMX_PREFER_STATIC_LIBS=ON` to work correctly for MKL. Note that this wasn't working correctly with `-qmkl` (prior to !2935) either. If we don't care about that, but do care about static linking support, it is sufficient to just take the start/end-group part of the 2nd commit.