Test C++17 standard library more robustly
Sometimes the std::filesystem library is in a separate library, which
can now be found, even if it comes from libc++.

The existing code needs to be re-ordered so that if it is appropriate
to look for the standard library (and then perhaps a separate
filesystem library) the search for the filesystem library is done
after handling GMX_GPLUSPLUS_PATH. GROMACS now checks that the compiler
accepts --gcc-toolchain flag before using it.

Avoided the anti-pattern of re-using cache variables. Updated various
comments. Removed support for the Intel classic compiler, as GROMACS
already does not support that.

Closes #4718
Closes #4728