Move enforced rotation regression tests to unit testing framework
The tests for the rotational pulling module still live in the regressiontests repository.
This change implements the same tests in the main GROMACS repository using GTest functionality.

When this change is accepted, the corresponding tests in the regressiontests repository can be deleted (in directories rotation/*); see gromacs-regressiontests#7.

This is a preparation to convert the enforced rotation module to use the IForceProvider interface,
which will clean up the MD loop.