Increase and check for limits of mdrun output files
The checkpoint and trr file store the number of elements in vectors
as an int. This put the limit on #atoms at max_int/3. Now we cast
from unsigned int to int, which doubles these limits.
There is now a check in mdrun for these limits so the user gets
a clear error message when limits are exceeded.

Removed unused nptr argument from doVectorLow() in checkpoint.cpp.

Refs #4627