Make a dependency explicit
gmx_pme_send_coordinates passes localXReadyOnDevice event pointer,
which must be valid when the coordinates should go from this PP rank's
GPU to a PME-rank's GPU. So the logic that sets the pointer should
have an explicit dependency on pmeSendCoordinatesFromGpu.

Currently we only do that transfer from this rank's GPU when
coordinate buffer ops are also happening on this rank's GPU, which
shares the same need for the event pointer to be valid, so there was
no actual problem. There would be a problem if the situation needed
to change.

Having done so, minimized the scope of reinitGpuPpPmeComms.