Remove FindFilesystem.cmake
This is a partial revert of 810c238902 because there's no known case
where it is needed. MSVC and gcc are either recent enough or not. A
clang-based compiler can be configured to use an appropriate libstdc++
installed somewhere (and that can even find libstdc++fs when forced to
use gcc8), or taught how to do so via other mechanisms such
as setting --gcc-toolchain or GMX_GPLUSPLUS_PATH.

Minor improvements to docs and logic for EXTRA_MESSAGE.

Refs #4728
Fixes #4796