Update bundled GoogleTest to 1.13.0
GoogleTest recommends updating as often as possible, but
it has been almost two years since the last update, so
living-at-HEAD approach is not working for us. Instead,
we update here to the latest tagged release (January 2023).
If later we need to use an untagged commit, that can still
be easily done.