SYCL/CMake: remove check for HIPSYCL_SYCLCC_EXTRA_COMPILE_OPTIONS
This CMake option was added to hipSYCL almost a year ago and is
available in all supported hipSYCL versions (0.9.4). It's now safe to
assume that it's available and works as intended.

Moved its setting to the beginning of the file to improve behavior or
GMX_ALL_HIPSYCL_CMAKE_FLAGS_COPY.

Refs #4716