Tell users how to work around lacking OpenMP support on Apple
We have previously explicitly not issued fatal errors for
lacking OpenMP support, since by far the most common platform
where this happend was Apple where we anyway can't use GPUs.
However, instead of reverting to a warning there now seems to
be a functioning drop-in version of libomp distributed by the
R project (including both x86 and arm64 code) that we can
point them to.

Refs #4791