Minor improvements to gmx_wallcycle
Add documentation and where appropriate default initializers for
fields of gmx_wallcycle.

This orders the little-used fields towards the end of the
gmx_wallcycle object. This help a little with cache locality in the
frequent case that we are not debugging wallcycle counting.

Moved debug helper functions to be member functions