Fix portability/reproducibility of erfinv
The tabulated normal distribution used to generate
velocities in grompp relies on erfinv(), which
we previously implemented with Newton-Raphson
iterations. This worked well on x86 that uses
denormal fp numbers, but it loses accuracy on Arm.
This replaces the erfinv functions with more
accurate implementations that at least appear to
generate identical results on different platforms,
which means the generated TPR files should also
be identical. It also tightens the tolerances
for all math functions so we are more likely to
detect similar issues in the future, and adjust
other reference values that appear to have been
produced at lower accuracy.

Fixes #4824.