Separate derived work licensing description
It is not a good idea to use additions at the end of
COPYING since it confuses automatic detection of the
license, even if our meaning was to be generous and
offer anyone to reuse derived works under their
original licenses. To fix this once and for all, this
change separates the latter into its own
COPYING.derived file.