Fix incorrect stepWork.clearGpuFBufferEarly flag
Add ListedForcesGpu::updateHaveInteractions to be able to call
setupDomainLifetimeWorkload earlier. This fixes the initialization of
stepWork.clearGpuFBufferEarly which previously used the stale and
potentially incorrect domainWork.haveCpuLocalForceWork and as a result
the local GPU force buffer clearing
could be missed with GPU halo exchange when after a DD partitioning
CPU bonded interactions are not present whereas before there were.

Fixes #4858