Fix energy minimization segfault when enforced rotation is active
By passing the correct pointer to the enforced rotation data in all calls of `do_force()` (instead of a `nullptr` in certain locations), now one can also do energy minimization when enforced rotation is active. This may make sense with rotation rates of 0, thus keeping the rotation about an axis as is.

Fixes #4865