Fix creating a filename from uninitialized data
Yet another `std::filesystem` bug.

Previously, we were passing the same `char*` for filename and title,
and `get_w_conf` was filling the title and later used it as a filename
in error messages. Now, since it needs a `path`, we were converting
uninitialized buffer into a `std::string` and then into `path`.

Noticed in an unrelated `valgrind` run, not sure it ever caused a bug.