put_atoms_in_box: use std::floor instead of loops
This MR adds some sanity checks in `gromacs/src/gromacs/pbcutil/pbc.cpp`to make sure that `put_atoms_in_box` works correctly and re-implements the loop in `put_atoms_in_box` to avoid hanging in case of infinite values. 

The tests consist in checking (1) that the coordinates of a point inside the box remain unchanged, (2) that the coordinates of a point outside a cubic box are correctly modified, (3) that the coordinates of a point outside a triclinic box are correctly modified and (4) that the function does not hang in case the coordinates are bad numerical values (if gmx is built in release mode).

In `put_atoms_in_box`, the inner-most loop has been substituted by `std::floor` to avoid hanging. I have timed the test to make sure that the new implementation is at least as efficient as the original. 

With the new implementation, running `put_atoms_in_box` 1000 times (10000 points each time) takes 314.42ms on average, while the old implementation takes 359.33ms (tested on Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz).

Closes #4766

Closes #4766