SYCL: Avoid runtime overhead of constructing sycl::property_list
With AdaptiveCpp/hipSYCL, we were constructing a new property_list for
every submission, which was doing unnecessary dynamic memory
allocation. Switched to static const.

On MI50 with ROCm 5.6.0 and AdaptiveCpp with the instant submission patch,
this MR reduces "Launch * GPU" counters by ~5% for the ADH benchmark.

Suggested by Aksel Alpay.

Refs #4659.