Colvars MDModule Interface
Add Colvars library in src/external and implement an MDModule interface for Colvars

This module supports full integration of the Colvars input and of its restart data with the Gromacs
TPR and CPT files, respectively and is compatible with parallel and GPU versions. mdp options were
added to activate the feature.
Note that the current interface gathers the relevant atomic coordinates on one MPI rank, 
where all collective variables and their forces are computed.
Tests and documentation are added.