Increase tolerance for one continuation test
The conserved energy quantity for the largest system for
the mdrun continuation test, alanine_vacuo, already had
a higher tolerance for single precision. Now a uniform
increase is used. This needs a higher tolerance because
the conserved energy is relatively small due to the sum
of (negative) potential energy and (positive) kinetic energy.

Fixes #4725