Add verlet-buffer-pressure-tolerance mdp parameter
Add mdp parameter that goes into the Verlet buffer size calculation
in addition to the drift on the energy tolerance. This parameters
limits the error in the average pressure due to mising Lennard-Jones
interactions. The default value of 0.5 bar causes a small performance
regression, in particular when running on GPUs and in particular
for coarse-grained systems.

Fixes #4861

Closes #4861