Improve GPU-aware MPI detection
Previously, during actual runs, only CUDA-awareness was checked. Now,
we check the awareness depending on which backend is actually used.

No changes expected with CUDA, improved auto-detection for SYCL.

The GPU-awareness reported by the underlying library is not reliable
(it does not report whether the specific hardware is supported), but
that's the same as it was.

Refs #4638