Prefer L0 backend to SYCL with dpcpp and GPU-aware Intel MPI
Intel MPI supports GPU-aware operations only on buffers created in
contexts on devices with L0 backends to the SYCL runtime. So in the
default case that devices with multiple backends are visible,
GROMACS will be easier to use if it defaults to the only choice
of backend that can work.

Refs #4638