Fix ASAN errors in unit tests
We were saving local (stack) variable addresses into a static
`t_pargs` structure.

In unit tests, that lead to a `stack-use-after-return` error.

Only `gmx check` was causing an error. 
In `gmx helixorient`, the stack variables are also static,
so there was no bug there, but it's easier to avoid
the dangerous pattern, for consistency.