Don't find MPI when we aren't going to use it
It's not always convenient to set -DCMAKE_DISABLE_FIND_PACKAGE_MPI=on
when wanting to avoid finding MPI when that MPI is broken (e.g. spack
builds).  We should have logic in libgromacs that permits avoiding
finding MPI uselessly so that simple cases work simply.

Improved some logic in python packaging so that we find an MPI
library+compiler when GROMACS has not done so, and issue a warning
about MPI compiler mismatch only when both compilers are found,
do mismatch, and the warning was not issued previously.