Update Colvars library to version "2023-12-04"
This change contains an update of the Colvars library in `src/external` (in a single commit) and small edits to GROMACS files (in separate commits for easier review).

The GROMACS commits address the following issues:
- MSVC compilation warnings (see #4924)
- Propagate OpenMP build flags to the library when requested for GROMACS
- Disable initialization to zero of the initial step number for Colvars, allowing it to restart itself from the cpt file. Not supported yet is the case where `init-step` is set to a non-zero value in the MDP (i.e. the Colvars internal step counter will always start at zero, unless when reading a checkpoint).
- Track the version of the GROMACS-Colvars interface.

The message of the single commit for the Colvars library contains links to pull requests in the Colvars repo. A notable feature is the keyword `defaultInputStateFile` to allow loading previous Colvars data from text file in the few cases where a cpt cannot be read. If a CPT is read, that keyword is ignored to avoid inconsistencies in the expected input.