Find libcuda and libnvidia-ml as stub libraries
This is a WAR for finding the stub libraries which is not supported in legacy FindCUDA module. At runtime the real libraries are found automatically. Now the NVSHMEM and cuFFTMp builds do not depend on the driver library being installed.
Note that this WAR does not support cross compilation cases.

Addresses #4886