RISC-V: use RDTIME instead of RDCYCLE
This is a backport of efb640d8 (!4040) with release-notes added.

Fixes #4967.

Starting with Linux 6.6 [1], RDCYCLE is a privileged instruction on
RISC-V and can't be used directly from userland. There is a sysctl
option to change that as a transition period, but it will eventually
disappear.

Use RDTIME instead, which while less accurate has the advantage of being
synchronized between CPU (and thus monotonic) and of constant frequency.
It is somehow similar to the virtual timer that is used by GROMACS on
Aarch64.

NB: it might be better to use the RDTIME pseudo instruction instead of
using the CSRRS instruction with the CSR addresses, but there might be
a reason for not using it.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=cc4c07c89aada16229084eeb93895c95b7eabaa3