Fixed usage of incomplete classes in containers
Necessary for C++20 compatibility; doesn't change anything with C++17 as far as I can see. The standard requires that a template type for a class be complete before any member of the resulting specialization is referenced. Thus various inline implementations have been moved to a source file, so that T will be complete by the time those specializations are referenced. None of those functions now inlined are important for mdrun performance.