Add CMake handling to manage compilation with HIP compiler
This change adds the initial CMake support to handle detection of the
ROCm toolkit and set up GROMACS compilation with it.

The general CMake code is very similar to the one for the CUDA backend,
and takes care of compiler file checking, compiler option setup and
target configuration.

No actual code is added here, the only code changes are for ensuring
that the detected new backend is not used to run any code that is not
implemented yet and will be reverted in following changes that add the
actual abstraction layer and later the kernels.

Addresses #4949