Fix crash in NbnxmSetupTest
A device pointer was not initialized to `nullptr` in `gpu_plist`.
In normal MD run, we called `nonbonded_verlet_t::constructPairlist` later,
which correctly initialized the buffer. But in the test, we were
immediately trying to free the buffer in `gpu_free`, passing the
uninitialized value to `cudaFree`.

The uninitialized buffer was added in 2023, but we only started freeing it
in 2024 (495c91159fd), so the problem only affects GROMACS 2024.

Fix proposed by @mdoijade in !3779

Co-authored-by: Mahesh Doijade <mdoijade@nvidia.com>

Closes #4888