Add NBNxM SIMD EnergyAccumulator class
The accumulation of energies in the NBNxM SIMD kernels is now handled
by the EnergyAccumulator class. The removes a lot of complexity from
the outer and inner loops of the kernel. It also makes the complex SIMD
energy group pair buffer handling an internal matter of this class.

The energy group per cluster storage is moved into the friend class
EnergyGroupsPerCluster. Thus the details of the packing of the energy
groups are completely hidden from the rest of the code.

This change is only refactoring, apart from adding tests.