Separate scheduling events for X and F GPU halo exchanges
In extreme run conditions, the tasks can become so out of sync that
there is a race condition on the event used to schedule
synchronization of X and F GPU halo exchanges, causing a seg fault in
cudaStreamWaitEvent. This fixes the issue by introducing separate
events for X and F halo scheduling.

Fixes #5024