Fix backup for Colvars output files
Makes the backup convention for Colvars output files consistent with other GROMACS output files (i.e. multiple backups). Currently, only a single backup file with either the `.old` or `.BAK` suffix appended is being produced, which is a convention inherited from NAMD.  See #5071

The revised interface code appends `_prev` to the Colvars state file's name (because is usually written at the same time as a checkpoint), and uses `make_backup()` otherwise.

This change may be cherry-picked on top of `main`.

A patch version number is also introduced for the Colvars library, to distinguish it from the version originally bundled with GROMACS 2024.0. The Colvars patch number is currently set to 2, because the GROMACS 2024.2 is being targeted here. The patch version does not need to be cherry-picked onto `main`, because it will be eventually reset when new Colvars features are added.