Used GoogleTest skip and failure mechanisms to manage NbnxmTests
Now we don't need to set a macro and recompile in order to safely
generate reference data. We also get a descriptive message in several
more cases where the test case needs to be skipped.

To do this, exposed the getter for the global mode for handling
reference data, which seemed better than trying to let the test case
push requirements onto TestReferenceData objects. Changed its name to
conform with style for getters.

Refs #5027