Grompp check for dihedral type 3 (RB or Fourier) coefficients summing to zero
Adds check to `grompp`, and associated unit test, to verify that the sum of the Ryckaert-Bellemans or Fourier coefficient is close to zero (absolute tolerance 0.01). Deviation from this has caused #4075 on our shipped OPLSaa.

One warning is emitted regardless of the number of instance of incorrect sum, to allow for `maxwarn` overriding, if it is somehow a legitimate use case.

Unexpectedly, and thanks to @al42and pushing for a `grompp` test, one of the test system in the `simulationdatabase` was found to trigger this check, `nonanol_vacuo`, with an error of 1.6736.

`nonanol_vacuo.top`
```
[ dihedrals ]
27	26	29	30	3	0.69873	2.09618	0.00000	-2.79491	0.00000	0.00000	0.69873	2.09618	0.00000	-2.79491	0.00000	0.00000
28	26	29	30	3	0.69873	2.09618	0.00000	-2.79491	0.00000	0.00000	0.69873	2.09618	0.00000	-2.79491	0.00000	0.00000
; offending FFParams:
1	5	8	11	3	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000
5	8	11	14	3	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000
8	11	14	17	3	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000
11	14	17	20	3	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000
14	17	20	23	3	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000
17	20	23	26	3	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000	3.68192	3.09616	-2.09200	-3.01248	0.00000	0.00000
20	23	26	29	3	0.65270	1.95811	0.00000	-2.61082	0.00000	0.00000	0.65270	1.95811	0.00000	-2.61082	0.00000	0.00000

```

This topology was created in 2015 in 2b600dc4c and is used in rerun and continuation tests (including checks of FEP on reruns). Since those are comparative/constistency tests to a reference trajectory, whether the parameters are physically correct is not that important.

Therefore a few questions:
- Are there legitimate cases where the RB dihedral coefficients can sum to something not zero?
- Can someone give some context on how/why this test topology was created?
- Opinion on how to resolve: removing this test system?  `-maxwarn`ing it away? Ideally, correcting the dihedral based on actual data, but if the tests using it are not about the actual physics, arbitrarily modifying the coefficient is also possible

Fixes #4253

Closes #4253 and #5009