Add missing synchronization before buffer reallocation
On NS steps, the force buffer was reallocated without ensuring that the
force clearing operation from the previous step completed. This led to
crashes with AdaptiveCpp task graph cashing enabled.

We add the missing synchronization here to ensure that there are no
pending operations in the Local Stream before we try to do any local
buffer reallocations.

Fixes #5078.