Clean up logic of gmx xpm2ps
The former code has several different outcomes according to whether
combining is in use, possibly with or without sometimes-optional
filesystem paths for XPM and EPS files.

Previously if ecombine was active and not ecHalves, an EPS file was
given, and an XPM file was not given, write_combined_matrix() would
try to open the missing XPM file and fail handling a nullptr. Now it
correctly requires the XPM file and the user gets a useful diagnostic.

Previously if the user gave neither of the optional paths, then
depending on ecombine one of the options was actually required and the
user got an explicit message from eg. ftp2fn. This is still true, but
expressed closer to the point of use so that we never have a variable
whose type is ambiguous about whether it holds a value when the
surrounding logic is clear that it must hold a value. Doing it this
way is easier to refactor automatically in future.

Also don't re-use the fn variable for the -f2 option.