Change Gitlab CI to use new GPU config Gitlab-runners
This changes our gitlab configuration to use a new setup with 
separate gitlab-runners for each type (and quantity) of GPUs, which
enables us to use default runners instead of hacking our own. It also
introduces variables that make it possible to skip all GPU runs with
a small change, if we e.g. briefly want to rely on default shared
gitlab runners in the future (like the past two weeks). For 
consistency, all inherited rules have been changed to references.

This is unfortunately a somewhat large change for the release branch,
but it is limited to the Gitlab-runner CI infrastructure and the change
per se thus won't affect any code.