Improve handling of FE init-lambda
There are two ways to initialize free-energy lambda values:
the old way with init_lambda and with lambda states. This is bug prone.
The change renaming init_lambda to init_lambda_without_states to
minimize changes for misinterpretation. It also adds an initLambda()
method to t_lambda to hide these two different ways.

Refs #4941