Add missing sqrt in Verlet buffer RMS inv.mass
The RMS inverse mass calculation for the Verlet buffer estimate was missing a sqrt. For systems with most masses much larger than 1, typically coarser-grained simulations, this could lead to particles incorrectly being flagged as constrained. This caused NaN drift estimates, which in turn lead to the Verlet buffer being set to zero.

This change adds the missing sqrt. It also add a check on the constraint length for detecting constrained atoms; this would also have avoids the NaNs. And finally, a release assert such that NaNs lead to mdrun quitting instead of producing unreliable results or crashing.

Fixes #5098