CUDA: Fix __launch_bounds__ for sm_87
We don't build sm_87 with NVCC CUDA, but we build it with Clang CUDA.
Previously, unsupported values were used, causing ptxas to emit a warning:

`Value of threads per SM for entry [...] is out of range. .minnctapersm will be ignored`

Older GROMACS versions also affected, but supporting Jetson Orin
is of low enough priority not to backport this patch.