The number of averaging blocks in einstein_visco is now chosen by the user
The number is averaging blocks used when computing viscosity using Einstein formula was hard-coded to 4. Now the user can decide the number of temporal chunks in which the integral of the formula is split.
Essentially, a new input variable has been added to `einstein_visco()`, `gmx_energy.cpp`. The value of the variable is set by the user via a new `-einstein_blocks` flag of `gmx energy`.

Partially solves #5114