Fix bugs when a PME spread pipeline is empty
This would only happen in a simulation with non-uniform density or a
test, but the code should handle it.

In a CUDA build with thread-MPI, no coordinate-send event will be
sent, so no receive should be posted, nor waited upon, and no kernel
should be launched.

In a GPU-aware MPI build, no coordinate receive is posted, so none
should be sent, and no kernel should be launched.