SYCL: Update the ROCm directory layout
Starting with ROCm 5.2, the file organization was changed to conform to
Linux Filesystem Hierarchy Standard (FHS). The old layout was supported
(via symlinks) in 5.x branch, but completely removed in ROCm 6. This
mainly affected builds with rocFFT.

Note that we previously used a mix of the old and the new layouts.

Also improve setting `ACPP_ROCM_PATH` by forcing it and indicating that
it is a directory, not a file.

See: https://rocm.docs.amd.com/en/docs-6.1.0/conceptual/file-reorg.html