Generalize and centralize helpers for naming tests and reference data
Now that we have a several different use cases, it makes sense to stop
copy-pasting and make it easier to add new tests that are only as
complex as they need to be.

If a parameterized test is defined in terms of a tuple then we can
also declare tuples of formaters that respectively map those values to
string fragments that can be combined to form either the test name or
the reference data name. This lets us write standardized helper code
for these common cases.

Unit tests for new test functionality

Refs #5027