Fix various crashs when Colvars lib was not built
Issues arise because the stub implementation of Colvars was too light
and did not defined the Colvars custom mdp variables, preventing tools
like `gmx check` or `gmx dump` to read correctly a tpr file generated
with GROMACS-Colvars version.

Prevent also the creation of a tpr file if Colvars module is activated
whereas GROMACS was not compiled with Colvars.

Proper exit if a Colvars simulation is launched whereas GROMACS was not
compiled with Colvars.

Refs #5055