Use ArrayRef in checkSequence
Fixes #4067 

The `checkSequence` function has been overloaded with a version taking an `ArrayRef` argument. Unit tests have been added to ensure that it matches the original function. 

Moving forward it would be simple to update most (or all?) function calls to use the reference version, and possibly remove the original version. Would this be of interest?

Closes #4067