Refactor constraints unit tests
This prepares to add new units tests to cover detection of the
existence of triangle constraints in a topology.

Separated the test fixture from the verification helper class, because
the new tests would like to share the range of topologies tested via
the test parameterization, but don't need reference data against which
to verify.

Added helper function to the runner to describe the constraint
algorithm being tested, rather than handling raw strings in order
to have algorithm-dependent test tolerances

Refs #5123