Fix calcQuasiHarmonicEntropy() in thermochemistry.cpp
Fixes #5041 

Fixes a calculation error in eigval_to_frequency() function, which now requires the temperature to be passed as a parameter. 

In order to pass temperature as a parameter, I had to add it to the temperature as an argument to the eigval_to_frequency() function as well as calcZeroPointEnergy(), as well as all instances that call these functions.