NNP Interface: Add Libtorch support to CMake
This is the first of several MRs related to the implementation of an interface for neural network potentials described in #5039. It adds capabilities to the CMake build system to link Gromacs against an external installation of a machine learning framework, as well as the `src/gromacs/applied_forces/nnpot` subdirectory. It's set up as a multichoice option, but currently only LibTorch, the C++ API of [PyTorch](https://github.com/pytorch/pytorch), is supported. It also allows to specify a path to an external shared library that implements a PyTorch [extension](https://pytorch.org/tutorials/advanced/cpp_extension.html).

Refs #5039