Implement NBNxM 1x1 gridding, search and reference nb-kernel
This change adds a new Nbnxm layout: 1x1. This actually does not use
our cluster pair layout, as all clusters consist of a single atom.
This layout can be more efficient when there are very few atom pairs
within the cut-off for each atom. Also this allows easy vectorization
of the innermost loop over the j-particles in the non-bonded kernel.

Refs #5150